refator(map layers): move leaflet layer junk into own directory
